+++ curlio/request.hpp
@@ -26,8 +26,8 @@ class Request
 public:
 	typedef boost::asio::any_io_executor executor_type;
 
-	Request()               = default;
-	Request(Request&& move) = delete;
+	Request() = default;
+	Request(Request&& move) noexcept;
 	~Request() noexcept;
 
 	void set_url(const char* url);
@@ -41,7 +41,7 @@ public:
 	auto async_write_some(const Const_buffer_sequence& buffers, Token&& token);
 	/// Getting the executor of an inactive request is undefined behavior.
 	boost::asio::any_io_executor get_executor() noexcept { return _data.lock()->executor; }
-	Request& operator=(Request&& move) = delete;
+	Request& operator=(Request&& move) noexcept;
 
 private:
 	friend Session;
@@ -53,9 +53,27 @@ private:
 	detail::Function<std::size_t(boost::system::error_code, void*, std::size_t)> _send_handler;
 
 	void _lazy_init();
+	void _hook(detail::Shared_data& data) noexcept;
 	static std::size_t _send_callback(void* data, std::size_t size, std::size_t count,
 	                                  void* self_pointer) noexcept;
 };
+
+inline Request::Request(Request&& move) noexcept
+{
+	std::swap(_owner, move._owner);
+	std::swap(_data, move._data);
+	std::swap(_request_headers, move._request_headers);
+	std::swap(_send_handler, move._send_handler);
+
+	if (_owner != nullptr) {
+		_hook(*_owner);
+	} else {
+		const auto data = _data.lock();
+		if (data != nullptr) {
+			_hook(*data);
+		}
+	}
+}
 
 inline Request::~Request() noexcept { curl_slist_free_all(_request_headers); }
 
@@ -135,6 +153,13 @@ inline auto Request::async_write_some(const Const_buffer_sequence& buffers, Toke
 	  token);
 }
 
+inline Request& Request::operator=(Request&& move) noexcept
+{
+	this->~Request();
+	new (this) Request{ std::move(move) };
+	return *this;
+}
+
 inline void Request::_lazy_init()
 {
 	if (!_data.expired()) {
@@ -142,11 +167,16 @@ inline void Request::_lazy_init()
 	}
 	_owner = std::make_shared<detail::Shared_data>();
 	_data  = _owner;
-	curl_easy_setopt(_owner->handle, CURLOPT_READFUNCTION, &Request::_send_callback);
-	curl_easy_setopt(_owner->handle, CURLOPT_READDATA, this);
+	_hook(*_owner);
+}
+
+inline void Request::_hook(detail::Shared_data& data) noexcept
+{
+	curl_easy_setopt(data.handle, CURLOPT_READFUNCTION, &Request::_send_callback);
+	curl_easy_setopt(data.handle, CURLOPT_READDATA, this);
 	// enable all encodings
-	curl_easy_setopt(_owner->handle, CURLOPT_ACCEPT_ENCODING, "");
-	_owner->request_finisher = [this] {
+	curl_easy_setopt(data.handle, CURLOPT_ACCEPT_ENCODING, "");
+	data.request_finisher = [this] {
 		if (_send_handler) {
 			_send_handler(boost::asio::error::eof, nullptr, 0);
 			_send_handler.reset();