+++ examples/playground.cpp
@@ -30,7 +30,7 @@ int main(int argc, char** argv)
 			  session.set_cookie_file("/tmp/cookme");
 			  for (int i = 0; i < 2; ++i) {
 				  curlio::Request req{};
-				  req.set_url("https://cloud.ayar.eu");
+				  req.set_url("https://example.com");
 				  // curl_easy_setopt(req.native_handle(), CURLOPT_VERBOSE, 1L);
 				  curl_easy_setopt(req.native_handle(), CURLOPT_FOLLOWLOCATION, 1L);
 				  // curl_easy_setopt(req.native_handle(), CURLOPT_USERAGENT, "curl/7.80.0");
@@ -45,12 +45,12 @@ int main(int argc, char** argv)
 				  // std::cout << "Done with artifical timeout\n";
 
 				  do {
-					  co_await resp->async_await_headers(use_awaitable);
+					  co_await resp.async_await_headers(use_awaitable);
 						std::cout << "=======RECEIVED HEADER======\n";
-				  } while (resp->is_redirect());
+				  } while (resp.is_redirect());
 				  std::cout << "Final headers received\n";
 
-				  co_await curlio::quick::async_read_all(*resp, use_awaitable);
+				  co_await curlio::quick::async_read_all(resp, use_awaitable);
 				  std::cout << "\nRead all data\n";
 
 				  // while (true) {
@@ -61,7 +61,7 @@ int main(int argc, char** argv)
 				  //   }
 				  //   // std::cout.write(buf, n);
 				  // }
-				  co_await resp->async_await_completion(use_awaitable);
+				  co_await resp.async_await_completion(use_awaitable);
 				  break;
 			  }
 		  } catch (const std::exception& e) {